Refactor Flask app.
