+++ bot/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bot = Blueprint('bot', __name__)
+
+from . import views
