+++ requirements.txt
@@ -3,6 +3,7 @@ certifi==2019.3.9
 chardet==3.0.4
 Click==7.0
 Flask==1.0.2
+Flask-Cors==3.0.7
 gunicorn==19.9.0
 idna==2.8
 itsdangerous==1.1.0
@@ -10,6 +11,7 @@ Jinja2==2.10
 MarkupSafe==1.1.1
 olefile==0.46
 requests==2.21.0
+six==1.12.0
 soupsieve==1.8
 urllib3==1.24.1
 Werkzeug==0.14.1
